Revert "refactor book controller"
This reverts commit 1935ed2bf1097ae6530f0f592f1f0022b5e3724c.
